% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tools.R
\name{ComputeEmpiricCramer}
\alias{ComputeEmpiricCramer}
\title{Computation of the Empiric Cramer'v.}
\usage{
ComputeEmpiricCramer(x)
}
\arguments{
\item{x}{a binary matrix.}
}
\value{
Return the matrix of the Empiric Cramer's V.
}
\description{
This function computes the Empiric Cramer's V for a binary data set.
}

