% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tools.R
\name{ComputeMvBinaryCramer}
\alias{ComputeMvBinaryCramer}
\title{Computation of the model Cramer'v.}
\usage{
ComputeMvBinaryCramer(results)
}
\arguments{
\item{results}{an instance of S4 class MvBinaryResult (provided by the function MvBinaryEstim)}
}
\value{
Return the matrix of the Empiric Cramer's V.
}
\description{
This function computes the model Cramer's V for a binary data set.
}

