% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MvBinaryEstim.R
\docType{data}
\name{MvBinaryExample}
\alias{MvBinaryExample}
\title{Simulated binary data: MvBinaryExample}
\format{A matrix with 400 observations on the 6 variables.}
\description{
The file MvBinaryExample.rda describes 400 individuals by 6 binary variables.
}
\details{
This data set has been simulated from the MvBinary model. The first three variables are dependent. The last three variables are dependent.
}
\examples{
data(MvBinaryExample)
}
\keyword{datasets}

