% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProbaPost.R
\name{MvBinaryProbaPost}
\alias{MvBinaryProbaPost}
\title{Computation of the model Cramer'v.}
\usage{
MvBinaryProbaPost(x, param)
}
\arguments{
\item{x}{a binary matrix.}

\item{param}{an instance of S4 class MvBinaryResult (provided by the function MvBinaryEstim)}
}
\value{
Return the logprobability for each row of matrix x conditionally on the model defined by param.
}
\description{
This function computes the model Cramer's V for a binary data set.
}

