% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MvBinaryEstim.R
\docType{data}
\name{plants}
\alias{plants}
\title{Real binary data: Plants}
\format{A matrix with 35583 observations on the 69 variables.}
\description{
The file plants.rda describes 35583 plants by indicating if they occur (1) or not (2) in 69 states of the Norht America.
}
\details{
This data set been extracted from the USA plants database, July 29, 2015.
}
\examples{
data(plants)
}
\keyword{datasets}

