% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,MvBinaryResult-method}
\title{Summary function.}
\usage{
\S4method{print}{MvBinaryResult}(x)
}
\arguments{
\item{x}{output object from \code{\link{MvBinaryEstim}}.}
}
\description{
This function prints the parameters resulting from \code{MvBinaryEstim}.
}

