% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,MvBinaryResult-method}
\title{Summary function.}
\usage{
\S4method{summary}{MvBinaryResult}(object)
}
\arguments{
\item{object}{output object from \code{\link{MvBinaryEstim}}.}
}
\description{
This function gives the summary of output from \code{MvBinaryEstim}.
}

