% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MvBinaryEstim.R
\docType{package}
\name{MvBinary-package}
\alias{MvBinary}
\alias{MvBinary-package}
\title{MvBinary a package for Multivariate Binary data}
\description{
MvBinary is a tool for fitting distribution on correlated multivariate binary data.
}
\details{
\tabular{ll}{
  Package: \tab MvBinary\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2015-07-29\cr
  License: \tab GPL-2\cr
  LazyLoad: \tab yes\cr
}
}
\examples{
rm(list=ls())
require(MvBinary)
data(MvBinaryExample)
test <- MvBinaryEstim(MvBinaryExample, nbcores = 3)
}
\author{
Author: Marbac M., and Sedki S.
}
\references{
Todo.
}
\keyword{package}

